large updates to functions and new function
created seqic_indicator_9

across all functions, provided a method where if the denominator is 0 in any case, then the performance calculated is NA.

updated is it normal to allow it to send information to the console if plotting is defaulted to FALSE due to passing more than one column.  Now, the function will put that info in a plotting behavior section instead of under normality tests.